pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "image_2_excel"
readme = "README.md"
authors = [{name = "Alex Spann", email = "alex.james.spann@gmail.com"}]
classifiers = [
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "License :: OSI Approved :: MIT License"
]
requires-python = ">=3.9"
dependencies = [
  "requests",
  "numpy",
  "xlsxwriter",
  "Pillow"
]
dynamic= ["version", "description"]

[tool.black]
edine-length = 88
target-version = ["py39"]

[tool.mypy]
python_version = "3.9"
show_error_codes = true
ignore_missing_imports = true